Splits smsprovider for different modules
