+++ app/src/dhisPlayServices/java/org/dhis2/utils/granularsync/GranularSyncModulePlayServices.kt
@@ -38,7 +38,7 @@ import org.dhis2.usescases.settings.models.ErrorModelMapper
 import org.hisp.dhis.android.core.D2
 
 @Module
-class GranularSyncModule(
+class GranularSyncModulePlayServices(
     private val context: Context,
     private val conflictType: SyncStatusDialog.ConflictType,
     private val recordUid: String,
@@ -74,7 +74,7 @@ class GranularSyncModule(
 
     @Provides
     fun smsSyncProvider(d2: D2): SMSSyncProvider {
-        return SMSSyncProviderImpl(
+        return SMSPlayServicesSyncProviderImpl(
             d2,
             conflictType,
             recordUid,
