+++ app/src/main/java/org/dhis2/utils/granularsync/GranularSyncPresenterImpl.kt
@@ -198,25 +198,23 @@ class GranularSyncPresenterImpl(
         return workManagerController.getWorkInfosForUniqueWorkLiveData(workName)
     }
 
-    override fun initSMSSync(): LiveData<List<SmsSendingService.SendingStatus>> {
-        statesList = ArrayList()
-        states = MutableLiveData()
-
+    override fun initSMSSync() {
         disposable.add(
             smsSyncProvider.getConvertTask()
                 .subscribeOn(Schedulers.io())
                 .observeOn(Schedulers.io())
-                .subscribeWith(
-                    smsSyncProvider.onConvertingObserver {
-                        updateStateList(it)
-                    }
+                .subscribe(
+                    { message -> view.openSmsApp(message) },
+                    { error -> Timber.e(error) }
                 )
         )
-
-        return states
     }
 
-    override fun sendSMS() {
+    override fun getGatewayNumber(): String {
+        return smsSyncProvider.getGatewayNumber()
+    }
+/*
+    /*override fun sendSMS() {
         disposable.add(
             smsSyncProvider.sendSms(
                 doOnNext = { sendingStatus: SmsSendingService.SendingStatus ->
@@ -236,29 +234,30 @@ class GranularSyncPresenterImpl(
                     }
                 )
         )
-    }
+    }*/
 
-    override fun onSmsNotAccepted() {
+   /* override fun onSmsNotAccepted() {
         updateStateList(
             smsSyncProvider.onSmsNotAccepted()
         )
-    }
+    }*/
 
-    private fun isLastSendingStateTheSame(sent: Int, total: Int): Boolean {
+    /*private fun isLastSendingStateTheSame(sent: Int, total: Int): Boolean {
         if (statesList.isEmpty()) return false
         val last = statesList[statesList.size - 1]
         return last.state == SmsSendingService.State.SENDING &&
             last.sent == sent &&
             last.total == total
-    }
+    }*/
 
-    private fun updateStateList(currentStatus: SmsSendingService.SendingStatus) {
+   /* private fun updateStateList(currentStatus: SmsSendingService.SendingStatus) {
         if (currentStatus.state != SmsSendingService.State.ERROR) {
             statesList.clear()
         }
         statesList.add(currentStatus)
         states.postValue(statesList)
-    }
+    }*/
+*/
 
     @VisibleForTesting
     fun getTitle(): Single<String> {
